requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
pandas==2.1.4
plotly==5.18.0
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1
psycopg2-binary==2.9.9
